VSM 0.4 pre-stable testing version
git-svn-id: https://vast.svn.sourceforge.net/svnroot/vast@41 e31416da-b748-0410-9b4e-cac2f8189b79

+++ dev/sim/VASTATEsim/simgame_arb.cpp
@@ -1,5 +1,25 @@
 
 /*
+ * VAST, a scalable peer-to-peer network for virtual environments
+ * Copyright (C) 2007-2008 Shao-Chen Chang (cscxcs at gmail.com)
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+ *
+ */
+
+/*
  *	SimGame_Arb.cpp  (Vastate Simulator Game _ Arbitrator class implementation)
  *
  *	
@@ -10,9 +30,10 @@
 
 char   simgame_arbitrator_logic::_ostr [MAXBUFFER_SIZE];
 errout simgame_arbitrator_logic::_eo;
+AttributeBuilder simgame_arbitrator_logic::atb;
 
 // callback - for authenaticing a newly joined peer
-bool simgame_arbitrator_logic::join_requested (id_t from_id, char *data, int sizes)
+bool simgame_arbitrator_logic::join_requested (VAST::id_t from_id, char *data, int sizes)
 {
 #ifdef VASTATESIM_DEBUG
     errout eo;
@@ -38,7 +59,7 @@ bool simgame_arbitrator_logic::join_requested (id_t from_id, char *data, int siz
     {
         char *auth = new char[JOIN_REQUEST_RANDOM_CHAR_LENGTH + 1];
         strncpy (auth, data, JOIN_REQUEST_RANDOM_CHAR_LENGTH);
-        _join_requesting [_storage->query (data, sizes)] = pair<id_t, char *> (from_id, auth);
+        _join_requesting [_storage->query (data, sizes)] = pair<VAST::id_t, char *> (from_id, auth);
     }
     else
     {
@@ -87,8 +108,98 @@ bool simgame_arbitrator_logic::query_returned (query_id_t query_id, char *data, 
     return true;
 }
 
+const std::vector<VAST::obj_id_t> &
+    simgame_arbitrator_logic::event_affected (VAST::id_t from_id, event &e)
+{
+    // return value
+    static std::vector<VAST::obj_id_t> ret;
+    ret.clear ();
+
+    int i_value;
+
+    // get sender object id
+	e.get (0, i_value);
+    VAST::id_t sender_object_id = (VAST::id_t) i_value;
+
+    // get sender object
+    object * sender_object = NULL;
+    if ((sender_object = get_object (sender_object_id)) == NULL)
+        return ret; 
+
+    // check for events
+    switch (e.type)
+    {
+        case SimGame::E_MOVE:
+            ret.push_back (sender_object->get_id ());
+            break;
+
+        case SimGame::E_ATTACK:
+        {
+            e.get (1, i_value);
+            VAST::id_t target_id = (VAST::id_t) i_value;
+
+            object *target_object;
+            if ((target_object = get_object (target_id)) == NULL)
+                break;
+            
+            if (target_object->get_pos ().dist (sender_object->get_pos ()) <= _para->VELOCITY * 2)
+                ret.push_back (target_id);
+        }
+        break;
+
+        case SimGame::E_BOMB:
+        {
+            int x, y, r;
+		    e.get (1, x);
+		    e.get (2, y);
+		    e.get (3, r);
+
+            Position dest = Position ((double) x, (double) y);
+
+#ifdef VASTATESIM_DEBUG
+            sprintf (_ostr, "[%d] INFO: arb_logic: bomb from [Peer%d] to (%d,%d) r %d" NEWLINE,
+            			_arbitrator->self->id, e.get_sender(), x,y,r);
+		    _eo.output(_ostr);
+#endif
+
+			map<VAST::id_t, object*>::iterator it;
+			for (it = _own_objects.begin(); it != _own_objects.end(); it ++)
+			{
+				object * obj = it->second;
+                if (dest.dist (obj->get_pos ()) <= (double) r)
+                    ret.push_back (obj->get_id ());
+            }
+		}
+		break;
+
+	    case SimGame::E_EAT:
+        {
+	        e.get (1, i_value);
+            VAST::id_t target_id = (VAST::id_t) i_value;
+
+            object * food;
+            if ((food = get_object (target_id)) == NULL)
+                break;
+
+            if (food->get_pos ().dist (sender_object->get_pos ()) <= _para->VELOCITY * 2)
+	        {
+                ret.push_back (sender_object_id);
+                ret.push_back (food->get_id ());
+	        }
+        }
+        break;
+
+        default:
+            sprintf (_ostr, "[%lu] ERROR: simgame_arbitrator_logic::event_affected (): received unknown event type (%d)", this->_arbitrator->self->id, 
+                            e.type);
+            _eo.output (_ostr);
+    }
+
+    return ret;
+}
+
 // callback -  for receiving a remote event (arbitrator only)
-bool simgame_arbitrator_logic::event_received (id_t from_id, event &e)
+bool simgame_arbitrator_logic::event_received (VAST::id_t from_id, event &e)
 {
     AttributeBuilder ab;
 #ifdef VASTATESIM_DEBUG
@@ -96,29 +207,19 @@ bool simgame_arbitrator_logic::event_received (id_t from_id, event &e)
 	char str[256];
 #endif
 
-    object * sender_object = NULL;
-
-	int sender_object_id;
+    int sender_object_id;
 	e.get (0, sender_object_id);
 
-    if (_own_objects.find (sender_object_id) == _own_objects.end ())
-    {
-#ifdef VASTATESIM_DEBUG
-        sprintf (str, "[%d] arbitrator_logic: received event of unknown sedner object %s\n",
-            (_arbitrator->self==NULL)?(0):(_arbitrator->self->id), ab.objectIDToString (sender_object_id).c_str ());
-        eo.output (str);
-#endif
+    // get sender object
+    object * sender_object = get_object (sender_object_id);
+    if (sender_object == NULL)
         return false;
-    }
-    else
-    {
-        sender_object = _own_objects[sender_object_id];
-    }
 
-	Position &sender_pos = sender_object->get_pos ();
+    Position &sender_pos = sender_object->get_pos ();
 
 	Position dest;
-	int x, y, r;
+	//double x, y, r;
+    float fx, fy, fr;
 	double dx, dy;
 	double ratio;
 	int target_id;
@@ -127,10 +228,10 @@ bool simgame_arbitrator_logic::event_received (id_t from_id, event &e)
 	switch (e.type)
 	{
 	case SimGame::E_MOVE:
-		e.get (1, x);
-		e.get (2, y);
-		dest.x = x;
-		dest.y = y;
+		e.get (1, fx);
+		e.get (2, fy);
+		dest.x = fx;
+		dest.y = fy;
 
 		dx = dest.x - sender_pos.x;
         dy = dest.y - sender_pos.y;
@@ -142,8 +243,8 @@ bool simgame_arbitrator_logic::event_received (id_t from_id, event &e)
         {
             // note that this may cause actual velocity to be less than desired
             // as any decimal places are dropped
-            dx = (long)((double)dx / ratio);
-            dy = (long)((double)dy / ratio);
+            dx = ((double)dx / ratio);
+            dy = ((double)dy / ratio);
         }
 
 		dest = sender_pos;
@@ -161,7 +262,7 @@ bool simgame_arbitrator_logic::event_received (id_t from_id, event &e)
 		break;
 
 	case SimGame::E_ATTACK:
-		e.get (1, (int)target_id);
+		e.get (1, target_id);
 
         if (_own_objects.find (target_id) == _own_objects.end ())
         {
@@ -201,25 +302,25 @@ bool simgame_arbitrator_logic::event_received (id_t from_id, event &e)
 		break;
 
 	case SimGame::E_BOMB:
-		e.get (1, (int)x);
-		e.get (2, (int)y);
-		e.get (3, (int)r);
+		e.get (1, fx);
+		e.get (2, fy);
+		e.get (3, fr);
 
-		dest.x = x;
-		dest.y = y;
+		dest.x = fx;
+		dest.y = fy;
 
 #ifdef VASTATESIM_DEBUG
 		sprintf (str, "[%d] arb_logic: bomb from [Peer%d] to (%d,%d) r %d" NEWLINE,
-			_arbitrator->self->id, e.get_sender(), x,y,r);
+			_arbitrator->self->id, e.get_sender(), (int) fx, (int) fy, (int) fr);
 		eo.output(str);
 #endif
 		
 		{
-			map<id_t, object*>::iterator it;
+			map<VAST::id_t, object*>::iterator it;
 			for (it = _own_objects.begin(); it != _own_objects.end(); it++)
 			{
 				target_object = (*it).second;
-				if (dest.dist (target_object->get_pos()) <= r)
+				if (dest.distsqr (target_object->get_pos()) <= (double) (fr*fr))
 				{
 					int userhp = ab.getPlayerHP(*target_object);
 					userhp -= _para->BOMB_POWER;
@@ -240,7 +341,7 @@ bool simgame_arbitrator_logic::event_received (id_t from_id, event &e)
 		break;
 
 	case SimGame::E_EAT:
-		e.get (1, (int)target_id);
+		e.get (1, target_id);
 		if (_own_objects.find (target_id) == _own_objects.end ())
 			break;
 
@@ -269,8 +370,8 @@ bool simgame_arbitrator_logic::event_received (id_t from_id, event &e)
 			else
 			{
 				//_reg->deletefood (target_object->get_id ());
-                _reg->dec_food ();
-				_arbitrator->delete_obj (target_object);
+				if (_arbitrator->delete_obj (target_object))
+                    _reg->dec_food ();
 			}
 		}
 #ifdef VASTATESIM_DEBUG
@@ -308,9 +409,10 @@ void simgame_arbitrator_logic::obj_created (object *obj, void *ref, size_t size)
 		//		statistics * sta = statistics::getInstance ();
 
 		// TODO: read datas from server
-		sprintf (ch, "Peer%d", obj->peer);
+		sprintf (ch, "Peer%lu", obj->peer);
 		//		bool buildresult = true;
-		bool buildresult = ab.buildPlayer (*obj, ch, rand () % _para->HP_MAX, _para->HP_MAX);
+		//bool buildresult = ab.buildPlayer (*obj, ch, rand () % _para->HP_MAX, _para->HP_MAX);
+		ab.buildPlayer (*obj, ch, rand () % _para->HP_MAX, _para->HP_MAX);
 		//		sta->objectChanged (1, player_object->get_id ());
 		
 #ifdef VASTATESIM_DEBUG
@@ -324,7 +426,7 @@ void simgame_arbitrator_logic::obj_created (object *obj, void *ref, size_t size)
 	{
 		if (ref == NULL)
 		{
-			sprintf (ch, "[%d] arb_logic: obj_created: Has no reference data." NEWLINE,
+			sprintf (ch, "[%lu] arb_logic: obj_created: Has no reference data.\n",
 				_arbitrator->self->id);
 			e.output(ch);
 		}
@@ -375,7 +477,8 @@ void simgame_arbitrator_logic::obj_discovered (object *obj)
 	_own_objects[obj->get_id()] = obj;
 
     // update god's object store
-    if (_arbitrator != NULL && _arbitrator->self != NULL)
+    if (_arbitrator != NULL && _arbitrator->self != NULL &&
+        _arbitrator->is_obj_owner (obj->get_id ()))
         _reg->update_object (_arbitrator->self->id, obj);
 }
 
@@ -393,12 +496,13 @@ void simgame_arbitrator_logic::obj_deleted (object *obj)
 	_own_objects.erase (obj->get_id ());
 
     // update god's object store
-    if (_arbitrator != NULL && _arbitrator->self != NULL)
+    if (_arbitrator != NULL && _arbitrator->self != NULL &&
+        _arbitrator->is_obj_owner (obj->get_id ()))
         _reg->delete_object (_arbitrator->self->id, obj);
 }
 
 // callback - by remote arbitrator to notify their object states have changed
-void simgame_arbitrator_logic::state_updated (id_t obj_id, int index, void *value, int length, version_t version)
+void simgame_arbitrator_logic::state_updated (VAST::id_t obj_id, int index, void *value, int length, version_t version)
 {
 #ifdef VASTATESIM_DEBUG
 	errout eo;
@@ -414,11 +518,12 @@ void simgame_arbitrator_logic::state_updated (id_t obj_id, int index, void *valu
         return ;
 
     // update god's object store
-    if (_arbitrator != NULL && _arbitrator->self != NULL)
+    if (_arbitrator != NULL && _arbitrator->self != NULL &&
+        _arbitrator->is_obj_owner (obj_id))
         _reg->update_object (_arbitrator->self->id, _own_objects[obj_id]);
 }
 
-void simgame_arbitrator_logic::pos_changed (id_t obj_id, Position &newpos, timestamp_t version)
+void simgame_arbitrator_logic::pos_changed (VAST::id_t obj_id, Position &newpos, timestamp_t version)
 {
 #ifdef VASTATESIM_DEBUG
     AttributeBuilder ab;
@@ -434,7 +539,8 @@ void simgame_arbitrator_logic::pos_changed (id_t obj_id, Position &newpos, times
         return ;
 
     // update god's object store
-    if (_arbitrator != NULL && _arbitrator->self != NULL)
+    if (_arbitrator != NULL && _arbitrator->self != NULL &&
+        _arbitrator->is_obj_owner (obj_id))
         _reg->update_object (_arbitrator->self->id, _own_objects[obj_id]);
 }
 
@@ -481,13 +587,13 @@ void simgame_arbitrator_logic::tick ()
     if (this_step < 0)
 		return;
 
-	map<id_t, Node> & arbs = _arbitrator->get_arbitrators ();
+	map<VAST::id_t, Node> & arbs = _arbitrator->get_arbitrators ();
 	for (int i = food_start_index[this_step]; i < food_start_index[this_step+1]; i++)
 	{
 		double food_dist = food_store[i].pos.dist (_arbitrator->self->pos);
 		bool nearest = true;
 
-		map<id_t, Node>::iterator it = arbs.begin ();
+		map<VAST::id_t, Node>::iterator it = arbs.begin ();
 		for (; it != arbs.end (); it ++)
 		{
 			if (food_dist > it->second.pos.dist (food_store[i].pos))
@@ -500,7 +606,8 @@ void simgame_arbitrator_logic::tick ()
 		if (nearest == true)
 		{
 			int type = 1;
-			object * newfood = _arbitrator->create_obj (food_store[i].pos, 0, (void *)&type);
+			//object * newfood = _arbitrator->create_obj (food_store[i].pos, 0, (void *)&type);
+			_arbitrator->create_obj (food_store[i].pos, 0, (void *)&type);
 		}
 	}
 
@@ -511,43 +618,78 @@ void simgame_arbitrator_logic::tick ()
     if (_detection_countdown <= 0)
     {
         // reset detection coundown
-        _detection_countdown = 10;
+        // TODO: change to normal value
+        _detection_countdown = 5;
 
-        // fix value for _bandwidth_usage in normal (_bandwidth_usage > 0 => -0.25
-        //                                                            < 0 =>  0.25
+        // fix value for _bandwidth_usage in normal (_bandwidth_usage > 0 => -0.1
+        //                                                            < 0 =>  0.1
         //                                                            = 0 =>  0
-        double fix_up = (_bandwidth_usage == 0 ? 0 : ((_bandwidth_usage>0?(-1):(1)) * 0.25));
-        
-        if (_reg->node_transmitted[_arbitrator->self->id] > THRESHOLD_BANDWIDTH)
-            fix_up = 1;
-        else if (_reg->node_transmitted[_arbitrator->self->id] < THRESHOLD_LOW_BANDWIDTH)
-            fix_up = -1;
+        double fix_up = 0.0;
+        if (_bandwidth_usage > 1)
+            fix_up = -1.0;
+        else if (_bandwidth_usage < -1)
+            fix_up = 1.0;
+        else 
+            fix_up = -1.0 * _bandwidth_usage;
+
+        if (_reg->node_transmitted[_arbitrator->self->id] >= THRESHOLD_BANDWIDTH * (_promoted ? RATIO_CAPABLE : 1.0))
+            fix_up = 1.0;
+        else if (_reg->node_transmitted[_arbitrator->self->id] <= THRESHOLD_LOW_BANDWIDTH * (_promoted ? RATIO_CAPABLE : 1.0))
+            fix_up = -1.0;
 
         _bandwidth_usage += fix_up;
 
-        if (_bandwidth_usage >= 3)
+#ifndef RUNNING_MODE_NO_AGGREGATOR
+#ifndef RUNNING_MODE_CLIENT_SERVER
+        // TODO: change to normal value
+        if (_bandwidth_usage >= 1.5)
         {
-            _arbitrator->overload ((int) _bandwidth_usage);
-            //fix_up = -3;
-            _bandwidth_usage = 1;
+            _arbitrator->overload ((int) (_bandwidth_usage + 0.5));
+            _bandwidth_usage = 0;
 
 #ifdef VASTATESIM_DEBUG
-            sprintf (_ostr, "[%d] call for overload help\r\n", _arbitrator->self->id);
+            sprintf (_ostr, "[%d] call for overload help\n", _arbitrator->self->id);
             _eo.output (_ostr);
 #endif
         }
-        else if (_bandwidth_usage <= -3)
+        else if (_bandwidth_usage <= -1.5)
         {
-            _arbitrator->underload ((int) _bandwidth_usage);
-            //fix_up = 3;
-            _bandwidth_usage = -1;
+            _arbitrator->underload ((int) (_bandwidth_usage - 0.5));
+            _bandwidth_usage = 0;
 
 #ifdef VASTATESIM_DEBUG
-            sprintf (_ostr, "[%d] call for underload help\r\n", _arbitrator->self->id);
+            sprintf (_ostr, "[%d] call for underload help\n", _arbitrator->self->id);
             _eo.output (_ostr);
 #endif
         }
+#endif /* RUNNING_MODE_CLIENT_SERVER */
+#endif /* RUNNING_MODE_NO_AGGREGATOR */
     }
+}
+
+
+VAST::id_t simgame_arbitrator_logic::_my_id () const
+{
+    if (_arbitrator == NULL || _arbitrator->self == NULL)
+        return NET_ID_UNASSIGNED;
+
+    return _arbitrator->self->id;
+}
+
+object *
+simgame_arbitrator_logic::get_object (obj_id_t object_id)
+{
+    if (_own_objects.find (object_id) == _own_objects.end ())
+    {
+#ifdef VASTATESIM_DEBUG
+        sprintf (_ostr, "[%d] arbitrator_logic: received event of unknown sedner object %s\n",
+            _my_id (), atb.objectIDToString (object_id).c_str ());
+        _eo.output (_ostr);
+#endif
+        return NULL;
+    }
+    else
+        return _own_objects[object_id];
 }
 
 
@@ -555,12 +697,25 @@ bool simgame_arbitrator_logic::getArbitratorInfo (arbitrator_info * arb_info)
 {
     if ((arb_info != NULL) && (_arbitrator->self != NULL))
 	{
-		arb_info->id = _arbitrator->self->id;
-		arb_info->pos = _arbitrator->self->pos;
-		return true;
+		arb_info->id      =  _arbitrator->self->id;
+		arb_info->pos     =  _arbitrator->self->pos;
+        arb_info->aoi     =  _arbitrator->self->aoi;
+        arb_info->joined  =  _arbitrator->is_joined ();
+        arb_info->is_aggr = _arbitrator->is_aggregator ();
+        arb_info->aoi_b   = (_arbitrator->is_aggregator ()) ? (_arbitrator->get_aggnode ()->aoi_b) : (0);
+        strcpy (arb_info->status, "");
+        if (_bandwidth_usage >= 1)
+            strcpy (arb_info->status, "+");
+        else if (_bandwidth_usage <= -1)
+            strcpy (arb_info->status, "-");
+
+        return true;
 	}
     arb_info->id = NET_ID_UNASSIGNED;
     arb_info->pos = Position (0, 0);
+    arb_info->aoi = arb_info->aoi_b = 0;
+    arb_info->joined = arb_info->is_aggr = false;
+
 	return false;
 }
 
@@ -589,13 +744,24 @@ const char * simgame_arbitrator_logic::toString ()
 	Node * n = _arbitrator->self;
     map<obj_id_t,bool> & obj_own = _arbitrator->get_owned_objs ();
 
+    if (_arbitrator->is_aggregator ())
+        sr << "[Aggregator]" << endl;
+    sr  << "Joined: " << _arbitrator->is_joined () << endl;
 	sr  << "ID: " << n->id << endl;
-	sr  << "AOI: " << n->aoi << endl;
+	sr  << "AOI: " << n->aoi;
+    // output second aoi if is aggregator
+    if (_arbitrator->is_aggregator ())
+    {
+        AggNode * ag = _arbitrator->get_aggnode ();
+        if (ag != NULL)
+            sr << "    AOI_b: " << ag->aoi_b;
+    }
+    sr << endl;
 	sr << "Pos: (" << n->pos.x << "," << n->pos.y << ")" << endl;
 	sr  << "Known objs: " << endl;
 
     // list peers first
-    for (map<id_t, object *>::iterator it2 = _own_objects.begin(); it2 != _own_objects.end(); it2 ++)
+    for (map<VAST::id_t, object *>::iterator it2 = _own_objects.begin(); it2 != _own_objects.end(); it2 ++)
     {
         if (it2->second->peer == 0)
             continue;
@@ -607,7 +773,7 @@ const char * simgame_arbitrator_logic::toString ()
     }
 
     // list normal object after
-    for (map<id_t, object *>::iterator it2 = _own_objects.begin(); it2 != _own_objects.end(); it2 ++)
+    for (map<VAST::id_t, object *>::iterator it2 = _own_objects.begin(); it2 != _own_objects.end(); it2 ++)
     {
         if (it2->second->peer != 0)
             continue;
@@ -618,7 +784,7 @@ const char * simgame_arbitrator_logic::toString ()
         sr << endl;
     }
 
-    sr << "Arbitrator info: " << endl;
+    //sr << "arbitrators: " << endl;
     sr << _arbitrator->to_string ();
 
 	*_info = sr.str();
@@ -645,6 +811,12 @@ void simgame_arbitrator_logic::register_interface (void *p_arb)
 void simgame_arbitrator_logic::register_storage (void * p_sto)
 {
     _storage = (storage *) p_sto;
+}
+
+bool 
+simgame_arbitrator_logic::get_info (int info_type, char* buffer, size_t & buffer_size)
+{
+    return _arbitrator->get_info (info_type, buffer, buffer_size);
 }
 
 // Recycle area /////////////////////////////////////////////////////////// 
@@ -704,3 +876,4 @@ void simgame_arbitrator_logic::peer_left (object * player_object)
 		}
 	}
 	*/
+
